remaining package migration squashed

+++ packages/Webkul/Product/src/Database/Migrations/2018_07_27_070011_create_product_attribute_values_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text('text_value')->nullable();
             $table->boolean('boolean_value')->nullable();
             $table->integer('integer_value')->nullable();
-            $table->double('float_value')->nullable();
+            $table->decimal('float_value', 12, 4)->nullable();
             $table->dateTime('datetime_value')->nullable();
             $table->date('date_value')->nullable();
             $table->json('json_value')->nullable();
